Changed CLAID.requestBatteryOptimizationExemption to show a warning on WearOS. On WearOS, requestBatteryOptimization via intent is not possible, because the intent is ignored on (multiple or all?) WearOS devices. Instead, the only option is to use adb to white the app package for battery optimization exemptions. The warning will show instructions on how to do that (i.e., the command that needs to be used in adb).
Changed MaximumPermissionsPerpetualService to also require permissions to show a notification. Without this permission, an ongoing notification
can not be shown to the user for a foreground service. This notification, however, is required to prevent that the process of an App goes into idle mode.
If it goes into idle, it can be killed after some while (2 days in my test).